Replace training-server with training-server-2
