--- molecule/omero-training-server-2/Dockerfile.j2
@@ -1,1 +1,0 @@
-../resources/Dockerfile.j2
