--- omero/training-server-2/files/prometheus-omero-additional-queries.yml
@@ -1,5 +1,0 @@
-# Additioal metric queries
-omero_countevents_actiontypes_total:
-  description: Number of events by type
-  query: 'SELECT COUNT(*), action, entityType From EventLog GROUP BY action, entityType'
-  labels: [action, entity]
