--- omero/training-server-2/letsencrypt.yml
@@ -1,81 +1,0 @@
-# Additional Nginx configuration including Let's Encrypt
-# Should be run when less than 30 days remain on the certificate
-
-- hosts: ome-outreach
-
-  pre_tasks:
-
-    - name: lentsencrypt challenge directory
-      become: yes
-      file:
-        path: /srv/www/letsencrypt/.well-known/
-        recurse: yes
-        state: directory
-
-  roles:
-
-    # Assume Nginx is already installednginx in certbot
-    # - role: ome.nginx
-
-    # If we're not using letsencrypt assume this is CI so use a self-signed cert
-    - role: ome.ssl_certificate
-      when: not (https_letsencrypt_enabled | default(False))
-
-    # Lets encrypt with automatic renewal
-    # This will stop nginx when the certificate is first created
-    # For renewals we configure Nginx to serve the challenge
-    - role: ome.certbot
-      become: yes
-      certbot_create_if_missing: yes
-      certbot_admin_email: "{{ letsencrypt_email }}"
-      certbot_domains:
-        - "{{ https_certificate_domain }}"
-      certbot_create_standalone_stop_services:
-        - nginx
-      certbot_auto_renew_deploy_hooks:
-        - systemctl reload nginx
-      certbot_auto_renew_args: --webroot --webroot-path /srv/www/letsencrypt/
-      # May be useful for testing:
-      #certbot_create_args: --test-cert
-      #certbot_auto_renew_args: --test-cert --force-renewal --webroot --webroot-path /srv/www/letsencrypt/
-      # WARNING: If you have a test certificate and need to convert it to a
-      # real certificate you may need to run
-      #   rm -rf /etc/letsencrypt/*
-      when: https_letsencrypt_enabled | default(False)
-
-  tasks:
-
-  - name: letsencrypt nginx configuration
-    become: yes
-    copy:
-      content: |
-        location /.well-known/ {
-            alias /srv/www/letsencrypt/.well-known/;
-        }
-
-        listen 443 ssl;
-        ssl_certificate {{ (https_letsencrypt_enabled | default(False)) | ternary(https_letsencrypt_cert_path + '/fullchain.pem', ssl_certificate_public_path) }};
-        ssl_certificate_key {{ (https_letsencrypt_enabled | default(False)) | ternary(https_letsencrypt_cert_path + '/privkey.pem', ssl_certificate_key_path) }};
-
-        if ($ssl_protocol = "") {
-          rewrite ^/(.*) https://$host/$1 permanent;
-        }
-      dest: /etc/nginx/conf.d-nested-includes/https.conf
-    notify:
-    - reload nginx
-
-  handlers:
-  - name: reload nginx
-    become: yes
-    service:
-      name: nginx
-      state: reloaded
-
-  vars:
-    # Must be defined somewhere
-    # https_certificate_domain:
-    letsencrypt_email: sysadmin@openmicroscopy.org
-    # This must match the expectations of certbot, do not change this:
-    https_letsencrypt_cert_path: "/etc/letsencrypt/live/{{ https_certificate_domain | default('localhost') }}"
-    # In production set this to True:
-    # https_lets_encrypt_enabled:
